Push final

+++ src/Controller/InscriptionController.php
@@ -29,10 +29,25 @@ class InscriptionController extends AbstractController
         $uneInscription->setSortie($sortie);
         $uneInscription->setDateInscription(new \DateTime());
 
-        $em->persist($uneInscription);
-        $em->flush();
-
-        $this->addFlash("success", "Vous vous êtes inscrit à la sortie ".$sortie->getNom()." !");
+        if($sortie->getDateLimiteInscription() > new \DateTime())
+        {
+            if(count($sortie->getInscriptions()) < $sortie->getNbInscriptionsMax())
+            {
+                $em->persist($uneInscription);
+                $em->flush();
+                $this->addFlash("success", "Vous vous êtes inscrit à la sortie ".$sortie->getNom()." !");
+            }
+            else
+            {
+                $this->addFlash("warning", "Nombres de places maximum atteint. Inscription à la sortie non effectuée.");
+                return $this->redirectToRoute("home");
+            }
+        }
+        else
+        {
+            $this->addFlash("warning", "La date limite d'inscription à été atteint. Inscription à la sortie non effectuée.");
+            return $this->redirectToRoute("home");
+        }
         return $this->redirectToRoute("sortie_afficher", $array);
     }
 
@@ -41,20 +56,24 @@ class InscriptionController extends AbstractController
      */
     public function seDesister(int $idSortie, EntityManagerInterface $em, UserInterface $user)
     {
-        $array = array(
-            "id" => $idSortie
-        );
-
         $repoSortie = $em->getRepository(Sortie::class);
         $sortie = $repoSortie->find($idSortie);
 
         $repoInscription = $em->getRepository(Inscription::class);
         $inscription = $repoInscription->findByIdSortieAndIdParticipant($sortie, $user);
 
-        $em->remove($inscription[0]);
-        $em->flush();
+        if($sortie->getDateHeureDebut() > new \DateTime())
+        {
+            $em->remove($inscription[0]);
+            $em->flush();
+            $this->addFlash("success", "Vous vous êtes désinscrit de la sortie !");
+        }
+        else
+        {
+            $this->addFlash("warning", "La sortie à déjà débuter, vous ne pouvez pas vous désinscrire !");
+            return $this->redirectToRoute("home");
+        }
 
-        $this->addFlash("success", "Vous vous êtes désinscrit de la sortie !");
-        return $this->redirectToRoute("home", $array);
+        return $this->redirectToRoute("home");
     }
 }
